code/backend/requirements.txt
Django==4.2.10
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-environ==0.11.2
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
django-filter==23.5
django-rest-auth==0.9.5
djangorestframework-simplejwt==5.3.0